Allow addon-local configuration
This allows an addon to have its own postcssOptions. For example, this can appear in an addon's index.js:

```js
var CssImport = require('postcss-import');
var CssNext = require('postcss-cssnext');

module.exports = {
  name: 'my-addon',
  options: {
    postcssOptions: {
      compile: {
        enabled: true,
        plugins: [
          { module: CssImport },
          { module: CssNext }
        ]
      }
    }
  }
};
```

Those options will remain properly scoped to the addon's own code and have no effect on the wider application (and may not be using ember-cli-postcss at all)

I also had to streamline the eslint configuration to get it working for me. This is now using the out-of-the-box eslint setup provided by the latest ember-cli. This removed several eslint plugins that weren't doing anything (given that we don't have any clientside code in the addon).